+++ docs/CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 2.0 (unreleased)
 ----------------
 
+- Switch to PyPDF2 which is maintained compared to pyPdf and can be used as
+  a drop-in replacement.
+  [saily]
+
 - Add travis-ci for Plone 4.1, Plone 4.2 and Plone 4.3.
   [saily]